Add files via upload
This is my first actual use of GitHub. I am a bit late to start using git hub for this project, so this update is from multiple times that I have worked on the Game. Currently I have developed a simple character controller that allows anything I attach the script to, to move. I have also created another script that rotates a game object to always point at the mouse. This will be useful later as I will use it to aim the players portal gun. Something that took me a while was actually creating an "original" character and creating a walking and idle animation for him. This was challenging for me as I struggle to use photoshop. I plan to further develop the walking animation so that it is smoother and does not actually move the character.